// libs/backend/common/otel/lib/src/type/otel-environment.type.ts
export interface OpenTelemetryEnvironment {
  NODE_ENV?: string;
  OTEL_ENABLED?: string;
  OTEL_EXPORTER_OTLP_ENDPOINT?: string;
  OTEL_EXPORTER_OTLP_HEADERS?: string;
  OTEL_EXPORTER_OTLP_METRICS_ENDPOINT?: string;
  OTEL_EXPORTER_OTLP_METRICS_HEADERS?: string;
  OTEL_EXPORTER_OTLP_TRACES_ENDPOINT?: string;
  OTEL_EXPORTER_OTLP_TRACES_HEADERS?: string;
  OTEL_METRIC_EXPORT_INTERVAL?: string;
  OTEL_SDK_DISABLED?: string;
  OTEL_SERVICE_VERSION?: string;
}